[PPSSPP] Fix invisible text on GUI Menus
When using hardware rendering (e.g. Direct3D, OpenGL etc) in PPSSPP, the text in the GUI menus is invisible.  This is likely a sync issue with the implementation of the PSP sceGUClear functions.

This PR changes clear_screen to use a loop to write pixels of the specified colour to uncached VRAM instead of using hardware functions to clear each frame on the GUI menus.  This works on both PPSSPP and real hardware, and shouldn't result in a noticeable performance hit when navigating menus.